# perf(build): migrate from webpack to rspack