rspack is pretty much a superset of webpack that's many times faster.

On this project it's over 3x faster:

```sh
$ hyperfine --warmup=3 'node node_modules/webpack-cli/bin/cli.js --mode production' 'node node_modules/@rspack/cli/bin/rspack.js --mode production'
Benchmark 1: node node_modules/webpack-cli/bin/cli.js --mode production
  Time (mean ± σ):      5.371 s ±  0.270 s    [User: 8.241 s, System: 0.402 s]
  Range (min … max):    4.962 s …  5.816 s    10 runs

Benchmark 2: node node_modules/@rspack/cli/bin/rspack.js --mode production
  Time (mean ± σ):      1.624 s ±  0.145 s    [User: 2.741 s, System: 0.289 s]
  Range (min … max):    1.481 s …  1.815 s    10 runs

Summary
  node node_modules/@rspack/cli/bin/rspack.js --mode production ran
    3.31 ± 0.34 times faster than node node_modules/webpack-cli/bin/cli.js --mode production
```