@justinmk 2025-08-19
> In December 2024, malicious versions of @rspack/core and @rspack/cli were published to npm due to a compromised token